requirements.txt
google-genai>=1.0.0
notion-client>=2.2.1
python-dotenv>=1.0.0
flask>=3.0.0
werkzeug>=3.0.0
